+++ packages/genesys/packages/react-components/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 3.7.4
+
+[Select] Add open prop to controll the visibility of the select
+[Package.json] Update `react-component-core`
+
 ## 3.7.3
 
 [withColor] Add defaultColor value